+++ android-accountmanager/src/main/java/com/andretietz/retroauth/RetroauthAndroid.kt
@@ -14,8 +14,8 @@ object RetroauthAndroid {
     return Retroauth.setup(
       retrofit,
       authenticator,
-      AndroidOwnerStorage(application),
-      AndroidCredentialStorage(application)
+      AndroidAccountManagerOwnerStorage(application),
+      AndroidAccountManagerCredentialStorage(application)
     )
   }
 }